requirements.txt
Django==1.6.7
South==1.0.0
django-cors-headers==0.12.0
django-jenkins==0.16.2
coverage==3.7.1
django-pipeline==1.3.20
djangorestframework==3.0.3
django-redis-cache==0.13.0
psycopg2==2.5.2
redis==2.10.3
hiredis==0.1.4

celery==3.1.15
django-celery==3.1.15

apache-libcloud==0.15.1
requests==2.2
python-dateutil==1.4.1
pytz==2014.7
python-ldap==2.4.13
Pillow==2.5.3

## ours
threepio==0.2.0
rtwo==0.2.7
caslib.py==2.1.0
chromogenic==0.1.2
jwt.py==0.1.0

## dev
flower==0.7.2 #TODO: Do we need a specific version here? -steve
pip-tools==0.3.5 #TODO: Do we need a specific version here? -steve
newrelic==2.34.0.29 # newrelic investigation.
ipython
ipdb
flake8
pep8
wheel

django-filter==0.9.1